Add the ability to get the focus from the asset
You can now get the focus from the asset using a new focus parameter. This parameter automatically sets the coordinates according to the focus set on the asset.

IMPORTANT: There is currently a problem with this implementation. Statamic saves the coordinates as percentage values. But ImageKit expects an absolute value in pixels.